Reorganized the osti.py module into an osti subpackage with individual modules for osti_record.py, osti_web_client.py and osti_web_parser.py
