+++ src/pds_doi_service/core/outputs/osti/__init__.py
@@ -1,0 +1,12 @@
+"""
+====
+OSTI
+====
+
+This package contains the OSTI-specific implementations for the abstract
+classes of the outputs package.
+"""
+
+from .osti_record import DOIOstiRecord
+from .osti_web_client import DOIOstiWebClient
+from .osti_web_parser import DOIOstiJsonWebParser, DOIOstiXmlWebParser
